extraction/requirements.txt
nltk===3.7
pymupdf===1.20.1
azure-ai-textanalytics==5.1.0